requirements.txt
numpy
matplotlib
Pillow
tensorflow
